brought the table component into the container. Deleted the separate table component
