+++ code-kata/src/containers/Catfish.js
@@ -2,7 +2,62 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { getCatfish } from '../actions/catfish-action';
-import ChannelCatTable from '../components/ChannelCatTable'
+import { withStyles } from 'material-ui/styles';
+import Table, { TableBody, TableCell, TableHead, TableRow } from 'material-ui/Table';
+import Typography from 'material-ui/Typography';
+import Paper from 'material-ui/Paper';
+
+
+
+const styles = theme => ({
+    root: {
+        width: '100%',
+        margin: '0 auto',
+        marginTop: theme.spacing.unit * 3,
+        overflowX: 'auto',
+        minWidth: 700,
+        maxWidth: 1400
+    },
+    title: {
+        textAlign: 'center',
+        color: '#666'
+    }
+});
+
+
+const createTable = (fish, title, classes) => {
+  
+  return (
+
+    <Paper className={classes.root}>
+      <Typography type="title">{title}</Typography>
+      <Table className={classes.table}>
+        <TableHead>
+          <TableRow>
+            <TableCell>Ranking</TableCell>
+            <TableCell>Angler</TableCell>
+            <TableCell>Weight</TableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {fish.map((cat, i)=> {
+            return (
+              <TableRow key={i}>
+                <TableCell>{i + 1}</TableCell>
+                <TableCell>{cat.angler}</TableCell>
+                <TableCell>{cat.weight}</TableCell>
+              </TableRow>
+            );
+          })}
+        </TableBody>
+      </Table>
+    </Paper>
+  )
+}
+
+
+
+
 
 class Catfish extends Component {
 
@@ -12,17 +67,19 @@ class Catfish extends Component {
 
 
   render () {
-    let { getCatfish } = this.props
+    let { flatHeads, channelCats, blueCats } = this.props.getCatfish;
+    let { classes } = this.props;
 
       return (
         <div>
-          <ChannelCatTable catfish={getCatfish.channelCats} title="Channel Cats"/>
-          <ChannelCatTable catfish={getCatfish.flatHeads} title="Flatheads"/>
-          <ChannelCatTable catfish={getCatfish.blueCats} title="Blue Catfish"/>
+          {flatHeads ? createTable(flatHeads, "Flatheads", classes) : null}
+          {channelCats ? createTable(channelCats, "Channel Cats", classes) : null}
+          {blueCats ? createTable(blueCats, "Blue Catfish", classes) : null}
         </div>
       )
   }
 }
+
 
 function mapStateToProps(state, props){
   return {
@@ -36,4 +93,6 @@ function matchDispatchToProps(dispatch){
   }
 }
 
-export default connect(mapStateToProps, matchDispatchToProps)(Catfish);
+const mapping = connect(mapStateToProps, matchDispatchToProps)(Catfish)
+
+export default withStyles(styles)(mapping);
